create base class for signing screens
