+++ app/lib/screens/signing/sign_with_text.dart
@@ -1,13 +1,8 @@
-import 'dart:convert';
-
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:threebotlogin/helpers/logger.dart';
-import 'package:threebotlogin/models/wallet.dart';
 import 'package:threebotlogin/providers/wallets_provider.dart';
-import 'package:threebotlogin/services/signing_service.dart';
-import 'package:http/http.dart' as http;
+import 'package:threebotlogin/screens/signing/signing_mixin.dart';
 
 class SignWithTextScreen extends ConsumerStatefulWidget {
   const SignWithTextScreen({super.key});
@@ -16,335 +11,100 @@ class SignWithTextScreen extends ConsumerStatefulWidget {
   ConsumerState<SignWithTextScreen> createState() => _SignWithTextScreenState();
 }
 
-class _SignWithTextScreenState extends ConsumerState<SignWithTextScreen> {
-  final TextEditingController _textController = TextEditingController();
-  Wallet? selectedWallet;
-  String? signedData;
-  bool isLoading = false;
+class _SignWithTextScreenState extends ConsumerState<SignWithTextScreen>
+    with SigningMixin {
   String? textError;
-  String? walletError;
-  final TextEditingController _destUrlController = TextEditingController();
-  String? destUrlError;
 
   @override
-  void initState() {
-    super.initState();
-  }
-
-  @override
-  void dispose() {
-    _textController.dispose();
-    _destUrlController.dispose();
-    super.dispose();
-  }
-
-  void _validateInputs() {
+  void validateInputs() {
     setState(() {
-      if (_textController.text.isEmpty) {
+      if (textController.text.isEmpty) {
         textError = 'Please enter text to sign';
       } else {
         textError = null;
       }
-
-      if (selectedWallet == null) {
-        walletError = 'Please select a wallet';
-      } else {
-        walletError = null;
-      }
     });
+    validateWallet();
   }
 
-  bool _validateDestUrl() {
-    if (_destUrlController.text.isEmpty) {
-      return true;
-    }
-
-    String url = _destUrlController.text.trim();
-    try {
-      final uri = Uri.parse(url);
-      if (!uri.isScheme('http') && !uri.isScheme('https')) {
-        setState(() {
-          destUrlError = 'URL must start with http:// or https://';
-        });
-        return false;
-      }
-
-      if (!uri.hasAuthority) {
-        setState(() {
-          destUrlError = 'Invalid URL format';
-        });
-        return false;
-      }
-
-      setState(() {
-        destUrlError = null;
-      });
-      return true;
-    } catch (e) {
-      setState(() {
-        destUrlError = 'Invalid URL format';
-      });
-      return false;
-    }
-  }
-
-  Future<void> _signText() async {
-    _validateInputs();
-    if (textError != null || walletError != null) {
-      return;
-    }
-    if (!_validateDestUrl()) {
-      return;
-    }
-
+  @override
+  void setSigningError(String message) {
     setState(() {
-      isLoading = true;
-      signedData = null;
+      textError = message;
     });
-
-    try {
-      signedData = await md5Sign(
-        data: _textController.text,
-        walletSecretSeed: selectedWallet!.tfchainSecret,
-      );
-      if (_destUrlController.text.isNotEmpty) {
-        await sendSignedData(_destUrlController.text.trim(), signedData!);
-      }
-    } catch (e) {
-      if (mounted) {
-        setState(() {
-          logger.e('Failed to sign text: $e');
-          textError = 'Failed to sign text';
-        });
-      }
-    } finally {
-      if (mounted) {
-        setState(() {
-          isLoading = false;
-        });
-      }
-    }
-  }
-
-  Future<void> sendSignedData(String destUrl, String signature) async {
-    try {
-      final response = await http.post(
-        Uri.parse(destUrl),
-        headers: {'Content-Type': 'application/json'},
-        body: json.encode({'signature': signature}),
-      );
-
-      if (response.statusCode != 200) {
-        throw Exception('Failed to send signature to destination');
-      }
-
-      if (mounted) {
-        ScaffoldMessenger.of(context).showSnackBar(
-          SnackBar(
-            content: Text('Signature sent successfully',
-                style: Theme.of(context).textTheme.bodyMedium?.copyWith(
-                    color: Theme.of(context).colorScheme.onPrimaryContainer)),
-            backgroundColor: Theme.of(context).colorScheme.primary,
-          ),
-        );
-      }
-    } catch (e) {
-      logger.e('Error sending signature to destination: $e');
-      if (mounted) {
-        ScaffoldMessenger.of(context).showSnackBar(
-          SnackBar(
-            content: Text('Failed to send signature to destination',
-                style: Theme.of(context).textTheme.bodyMedium?.copyWith(
-                    color: Theme.of(context).colorScheme.onErrorContainer)),
-            backgroundColor: Theme.of(context).colorScheme.error,
-          ),
-        );
-      }
-    }
   }
 
   @override
   Widget build(BuildContext context) {
     final wallets = ref.watch(walletsNotifier);
-
     return Scaffold(
-      appBar: AppBar(
-        title: const Text('Sign Text'),
-      ),
-      body: SingleChildScrollView(
-        padding: const EdgeInsets.all(16),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.stretch,
-          children: [
-            Text(
-              'Enter Text to Sign',
-              style: Theme.of(context).textTheme.titleLarge?.copyWith(
-                    fontWeight: FontWeight.bold,
-                    color: Theme.of(context).colorScheme.onSurface,
-                  ),
-            ),
-            const SizedBox(height: 16),
-            TextField(
-              style: Theme.of(context).textTheme.bodyMedium?.copyWith(
-                    color: Theme.of(context).colorScheme.onSurface,
-                  ),
-              controller: _textController,
-              maxLines: 5,
-              onChanged: (value) {
-                if (textError != null) {
-                  _validateInputs();
-                }
-              },
-              decoration: InputDecoration(
-                hintText: 'Enter your text here...',
-                errorText: textError,
-                border: OutlineInputBorder(
-                  borderRadius: BorderRadius.circular(8),
-                ),
-              ),
-            ),
-            const SizedBox(height: 24),
-            Text(
-              'Select Wallet',
-              style: Theme.of(context).textTheme.titleLarge?.copyWith(
-                    fontWeight: FontWeight.bold,
-                    color: Theme.of(context).colorScheme.onSurface,
-                  ),
-            ),
-            const SizedBox(height: 16),
-            DropdownButtonFormField<Wallet>(
-              value: selectedWallet,
-              decoration: InputDecoration(
-                border: OutlineInputBorder(
-                  borderRadius: BorderRadius.circular(8),
-                ),
-                hintText: 'Select a wallet',
-                errorText: walletError,
-              ),
-              items: wallets.map((wallet) {
-                return DropdownMenuItem(
-                  value: wallet,
-                  child: Text(wallet.name,
-                      style: Theme.of(context).textTheme.bodyMedium?.copyWith(
-                            color: Theme.of(context).colorScheme.onSurface,
-                          )),
-                );
-              }).toList(),
-              onChanged: (Wallet? value) {
-                setState(() {
-                  selectedWallet = value;
-                  if (walletError != null) {
-                    _validateInputs();
-                  }
-                });
-              },
-            ),
-            const SizedBox(
-              height: 24,
-            ),
-            TextField(
-              controller: _destUrlController,
-              style: Theme.of(context).textTheme.bodyMedium?.copyWith(
-                    color: Theme.of(context).colorScheme.onSurface,
-                  ),
-              decoration: InputDecoration(
-                labelText: 'Destination URL (Optional)',
-                errorText: destUrlError,
-                hintText: 'https://example.com/api/signatures',
-                border: OutlineInputBorder(
-                  borderRadius: BorderRadius.circular(8),
-                ),
-              ),
-              onChanged: (value) {
-                if (value.isNotEmpty) {
-                  _validateDestUrl();
-                } else {
-                  setState(() {
-                    destUrlError = null;
-                  });
-                }
-              },
-            ),
-            const SizedBox(height: 32),
-            ElevatedButton(
-              onPressed: isLoading ? null : _signText,
-              style: ElevatedButton.styleFrom(
-                padding: const EdgeInsets.all(16),
-                shape: RoundedRectangleBorder(
-                  borderRadius: BorderRadius.circular(8),
-                ),
-              ),
-              child: isLoading
-                  ? const SizedBox(
-                      height: 20,
-                      width: 20,
-                      child: CircularProgressIndicator(strokeWidth: 2),
-                    )
-                  : Text('Sign',
-                      style: Theme.of(context).textTheme.titleLarge?.copyWith(
-                            color: Theme.of(context)
-                                .colorScheme
-                                .onPrimaryContainer,
-                            fontWeight: FontWeight.bold,
-                          )),
-            ),
-            if (signedData != null) ...[
-              const SizedBox(height: 24),
-              Row(
+        appBar: AppBar(
+          title: const Text('Sign Text'),
+        ),
+        body: SingleChildScrollView(
+            padding: const EdgeInsets.all(16),
+            child: Column(
+                crossAxisAlignment: CrossAxisAlignment.stretch,
                 children: [
                   Text(
-                    'Signed Data',
+                    'Enter Text to Sign',
                     style: Theme.of(context).textTheme.titleLarge?.copyWith(
                           fontWeight: FontWeight.bold,
                           color: Theme.of(context).colorScheme.onSurface,
                         ),
                   ),
-                  const SizedBox(width: 8),
-                  Text(
-                    '(hex encoded)',
-                    style: Theme.of(context).textTheme.bodySmall?.copyWith(
-                          color: Theme.of(context)
-                              .colorScheme
-                              .onSurface
-                              .withOpacity(0.6),
+                  const SizedBox(height: 16),
+                  TextField(
+                    style: Theme.of(context).textTheme.bodyMedium?.copyWith(
+                          color: Theme.of(context).colorScheme.onSurface,
                         ),
-                  ),
-                ],
-              ),
-              const SizedBox(height: 16),
-              Container(
-                padding: const EdgeInsets.all(16),
-                decoration: BoxDecoration(
-                  border: Border.all(
-                    color: Theme.of(context).colorScheme.outline,
-                  ),
-                  borderRadius: BorderRadius.circular(8),
-                ),
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    Text(signedData!,
-                        style: Theme.of(context).textTheme.bodyMedium!.copyWith(
-                              color: Theme.of(context).colorScheme.onSurface,
-                            )),
-                    const SizedBox(height: 8),
-                    TextButton.icon(
-                      icon: const Icon(Icons.copy),
-                      label: const Text('Copy'),
-                      onPressed: () {
-                        Clipboard.setData(ClipboardData(text: signedData!));
-                        ScaffoldMessenger.of(context).clearSnackBars();
-                        ScaffoldMessenger.of(context).showSnackBar(
-                            const SnackBar(content: Text('Copied!')));
-                      },
+                    controller: textController,
+                    maxLines: 5,
+                    onChanged: (value) {
+                      if (textError != null) {
+                        validateInputs();
+                      }
+                    },
+                    decoration: InputDecoration(
+                      hintText: 'Enter your text here...',
+                      errorText: textError,
+                      border: OutlineInputBorder(
+                        borderRadius: BorderRadius.circular(8),
+                      ),
                     ),
-                  ],
-                ),
-              ),
-            ],
-          ],
-        ),
-      ),
-    );
+                  ),
+                  const SizedBox(height: 24),
+                  buildWalletSelector(wallets),
+                  const SizedBox(height: 24),
+                  buildDestinationUrlField(),
+                  const SizedBox(height: 32),
+                  ElevatedButton(
+                    onPressed: isLoading ? null : signText,
+                    style: ElevatedButton.styleFrom(
+                      padding: const EdgeInsets.all(16),
+                      shape: RoundedRectangleBorder(
+                        borderRadius: BorderRadius.circular(8),
+                      ),
+                    ),
+                    child: isLoading
+                        ? const SizedBox(
+                            height: 20,
+                            width: 20,
+                            child: CircularProgressIndicator(strokeWidth: 2),
+                          )
+                        : Text('Sign',
+                            style: Theme.of(context)
+                                .textTheme
+                                .titleLarge
+                                ?.copyWith(
+                                  color: Theme.of(context)
+                                      .colorScheme
+                                      .onPrimaryContainer,
+                                  fontWeight: FontWeight.bold,
+                                )),
+                  ),
+                  const SizedBox(height: 24),
+                  buildSignedDataDisplay(),
+                ])));
   }
 }
